// src/app/ren/_lib/types.ts
export interface DivinationNums {
  num1: number;
  num2: number;
  num3: number;
}

export interface ManualNums {
  num1: number | null;
  num2: number | null;
  num3: number | null;
}

export interface DivinationResult {
  divination1: string;
  divination2: string;
  divination3: string;
  explain1: string;
  explain2: string;
  explain3: string;
}